add feature: detail product - change harga int to long - create OnItemClickListener - add some utils

+++ app/src/main/java/com/example/tokoonline/data/model/Produk.kt
@@ -13,7 +13,7 @@ data class Produk(
     @SerializedName("nama")
     var nama: String = "",
     @SerializedName("harga")
-    var harga: Int = 0,
+    var harga: Long = 0,
     @SerializedName("deskripsi")
     val deskripsi: String = "",
     @SerializedName("id_users")
